Added Coroutines, and added in UI elements

+++ COMP395 Project 1/Assets/Scripts/Customer.cs
@@ -14,7 +14,7 @@ public class Customer
         serviceTime = sTime;
     }
 
-    public float getID(){
+    public int getID(){
         return ID;
     }
     public float getArrivalTime(){
